refactor(canvas): finish chat component colocation
