+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMessages/ChatClarificationCard.tsx
@@ -1,9 +1,9 @@
 import { useId } from 'react';
-import { useI18n } from '../../i18n';
-import { isImeComposing } from '../../utils/ime';
-import { BotAvatarIcon } from '../icons';
-import { Button } from '../ui';
-import type { PendingClarification } from '../../types';
+import { useI18n } from '../../../i18n';
+import { isImeComposing } from '../../../utils/ime';
+import { BotAvatarIcon } from '../../icons';
+import { Button } from '../../ui';
+import type { PendingClarification } from '../../../types';
 
 interface ChatClarificationCardProps {
   pendingClarify: PendingClarification;
