+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody/RailToggleIcon.tsx
