+++ apps/canvas-workspace/src/renderer/src/components/chat/SessionBackBar/index.tsx
@@ -1,4 +1,4 @@
-import { useI18n } from '../../i18n';
+import { useI18n } from '../../../i18n';
 
 /** One entry in the session back-navigation stack (where a jump started). */
 export interface SessionBackEntry {
