+++ apps/canvas-workspace/src/renderer/src/components/chat/__tests__/ChatHeader.cross-scope.test.tsx
@@ -3,7 +3,7 @@ import { act, createRef } from 'react';
 import { createRoot } from 'react-dom/client';
 import { describe, expect, it, vi } from 'vitest';
 import { I18nProvider } from '../../../i18n';
-import { ChatHeader } from '../ChatHeader';
+import { ChatHeader } from '../ChatPanel/ChatHeader';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
